+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2.22 branch
 ### 2.21.1
 * `Html::youtube()` and `BBCode::youtube()` methods can handle the starting second (`start` get parameter);
+* `BBCode::youtube()` method removes unnecessary `<p>` tags around the code;
 * added `BBCode::hr()` method;
 * (re)added `HtmlHelper::shareaholic()` method.